Smooth Slopes Implemented
Custom curves now support having nonzero grades at either end.  Currently, the grade on each piece is selected by shift right clicking to up the grade by 1%, and shift left clicking to lower the grade by 1%. The current grade is outputted to the terminal.

Issues:
-There are some graphical glitches at the ends of slopes and where rails overlap.
-Floating point math sometimes causes flat tracks to destroy the blocks they are placed on and be very slightly sunk into the ground. (Most likely fixed by addding a small extra height to each rail)
-There's no feedback to the player what grade they have selected

Possible Expansion:
-Have the track blueprint/golden spike automatically take the grade from the piece of track they are used on. This is proably doable with onItemUse().

+++ src/main/java/cam72cam/immersiverailroading/items/ItemGoldenSpike.java
@@ -103,13 +103,13 @@ public class ItemGoldenSpike extends Item {
 		}
 		NBTTagCompound mainTag = stack.getTagCompound();
 		if(mainTag.hasKey("grade")) {
-			mainTag.setInteger("grade", mainTag.getInteger("grade") + 1);
+			mainTag.setFloat("grade", mainTag.getFloat("grade") + 1);
 		}
 		else {
-			mainTag.setInteger("grade", 1);
+			mainTag.setFloat("grade", 1);
 		}
 
-		System.out.println("Adjust spikePos up, now: " + mainTag.getInteger("grade"));
+		System.out.println("Adjust spikePos up, now: " + mainTag.getFloat("grade"));
 	}
 
 	private static void decrementGrade(ItemStack stack) {
@@ -119,16 +119,16 @@ public class ItemGoldenSpike extends Item {
 		}
 		NBTTagCompound mainTag = stack.getTagCompound();
 		if(mainTag.hasKey("grade")) {
-			mainTag.setInteger("grade", mainTag.getInteger("grade") - 1);
+			mainTag.setFloat("grade", mainTag.getFloat("grade") - 1);
 		}
 		else {
-			mainTag.setInteger("grade", -1);
+			mainTag.setFloat("grade", -1);
 		}
 
-		System.out.println("Adjust spikePos down, now: " + mainTag.getInteger("grade"));
+		System.out.println("Adjust spikePos down, now: " + mainTag.getFloat("grade"));
 	}
 
-	private int getGrade(ItemStack stack) {
+	private float getGrade(ItemStack stack) {
 		//I don't know if this null protection is necessary
 		if(stack.getTagCompound() == null) {
 			stack.setTagCompound(new NBTTagCompound());
@@ -136,7 +136,7 @@ public class ItemGoldenSpike extends Item {
 
 		NBTTagCompound mainTag = stack.getTagCompound();
 		if(mainTag.hasKey("grade")) {
-			return mainTag.getInteger("grade");
+			return mainTag.getFloat("grade");
 		}
 		else {
 			return 0;
